requirements.txt
Flask==3.0.2
PyJWT==2.8.0
bcrypt==4.1.2
pytest==8.0.2
requests==2.31.0
